Change tool registration so that everyting works on Windows.
The self-registration of tools would lead to crashes on Windows due to
initialization order issues. This is now replaced with explicit
registration of a list of tools in function register_all_tools().

The code has been written to create minimum compile-time depencencies,
so that tools still can be compiled in splendid isolation without
burdening the compilation unit of register_all_tools();